refactors views to be more consistent
