+++ src/lib/notifications/SuccessfullyJoinedStudyNotification.svelte
