+++ src/lib/notifications/SuccessfullyLeftStudyNotification.svelte
