+++ src/lib/views/terms/Content.svelte
