+++ src/routes/terms/index.svelte
@@ -3,7 +3,7 @@
 </script>
 
 <script lang="ts">
-  import TermsContent from "$lib/views/Terms.svelte";
+  import TermsContent from "$lib/views/terms/Content.svelte";
 </script>
 
 <TermsContent />
